BROKEN Refactor state handlers and update StateModel type
Moved state handlers to a new directory and updated StateModel to include an optional game logic function. Adjusted import paths across the codebase to reflect the new structure and added the game logic handler to gameStateModels.
